requirements.txt
streamlit
langchain
langchain-community
python-dotenv
requests
fpdf2
pypdf